Начальная логика работы боты

+++ src/main/java/ru/skypro/pets_home_bot/api_bot/service/PetUserService.java
@@ -23,5 +23,5 @@ public interface PetUserService {
      * @param chatIdPetUser - chatId усыновителя
      * @return - возвращается
      */
-    PetUser findByChatIdPetUser(int chatIdPetUser);
+    PetUser findByChatIdPetUser(long chatIdPetUser);
 }
